moved UI to components folder, and created route for saving messages in mongo backend only
