+++ client/src/components/Chat.css
@@ -1,0 +1,120 @@
+/* css of chat design */
+::-webkit-scrollbar {
+    width: 1em;
+  }
+  ::-webkit-scrollbar-track {
+    box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
+  }
+  ::-webkit-scrollbar-thumb {
+    background-color: darkgrey;
+    outline: 1px solid slategrey;
+    border-radius: 1em;
+  }
+  ::-webkit-input-placeholder {
+    text-align: center;
+    color: #bababa;
+    font-size: 16px;
+  }
+  :-moz-placeholder { /* Firefox 18- */
+    text-align: center;  
+    color: #bababa;
+    font-size: 16px;
+  }
+  ::-moz-placeholder {  /* Firefox 19+ */
+    text-align: center;  
+    color: #bababa;
+    font-size: 16px;
+  }
+  :-ms-input-placeholder {  
+    text-align: center; 
+    color: #bababa;
+    font-size: 16px;
+  }
+  html, body { height:100%; }
+  .bigCont{
+    width: 90%;
+    height:60%;
+  }
+  .chatDiv{
+    height: 60vh;
+    overflow-y: scroll;
+    margin: auto;
+    font-size: 12px;
+    background-color: #edeff1a3;
+    border-radius: 15px;
+    overflow-x: hidden;
+  }
+  .chatBox {
+      padding: 10px 10px;
+      display: flex;
+      justify-content: flex-end;
+      flex-direction: column;
+      align-self: flex-end;
+    color: #fff;
+  }
+  .inContainer {
+    display: inline-flex;
+  }
+  .inContainer.own {
+    flex-direction: row-reverse;
+  }
+  .inlineIcon {
+    width:20px;
+    object-fit: contain;
+  }
+  .ownDiv {
+      min-width: 60px;
+      max-width: 700px;
+      padding: 14px 18px;
+    margin: 6px 8px;
+      background-color: #5b5377;
+      border-radius: 16px 16px 0 16px;
+      border: 1px solid #443f56;
+  }
+  .otherDiv {
+      min-width: 60px;
+      max-width: 700px;
+      padding: 14px 18px;
+    margin: 6px 8px;
+      background-color: #6C8EA4;
+      border-radius: 16px 16px 16px 0;
+      border: 1px solid #54788e;
+  }
+  .own {
+      align-self: flex-end;
+  }
+  .other {
+      align-self: flex-start;
+  }
+  span.own,
+  span.other{
+    font-size: 14px;
+    color: grey;
+  }
+  .inputText{
+    width: 30%;
+  }
+  #formInputs{
+    width: 70%;
+    align-items: center;
+  }
+  .btnSend{
+    border-radius: 20px;
+  }
+  #nameForm {
+    opacity:    1; 
+    background: rgb(255, 255, 255); 
+    width:      100%;
+    height:     100%; 
+    z-index:    10;
+    top:        0; 
+    left:       0; 
+    position:   fixed; 
+  }
+  
+  #nameForm div{
+    margin-top: 15%;
+  }
+  #nameForm div input{
+    width: 40% !important;
+  }
