+++ client/src/components/LandingPage.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+class LandingPage extends Component {
+  render() {
+    return (
+      <h1>Landing</h1>
+    );
+  }
+}
+ 
+export default LandingPage;
