+++ package.json
@@ -26,6 +26,7 @@
     "multer": "^1.4.2",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0",
+    "react-router-dom": "^5.1.2",
     "request": "^2.88.0",
     "socket.io": "^2.3.0"
   },
